extracts signdocumentmanager from signingservice
