+++ src/Altinn.App.Core/Features/Signing/Models/Internal/SigningService.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Altinn.App.Core.Exceptions;
@@ -16,8 +15,6 @@ using Altinn.Platform.Register.Models;
 using Altinn.Platform.Storage.Interface.Models;
 using Microsoft.Extensions.Logging;
 using static Altinn.App.Core.Features.Signing.Models.Internal.Signee;
-using JsonException = Newtonsoft.Json.JsonException;
-using SigneeState = Altinn.App.Core.Features.Signing.Models.Internal.SigneeContextState;
 
 namespace Altinn.App.Core.Features.Signing.Models.Internal;
 
@@ -29,6 +26,7 @@ internal sealed class SigningService(
     IAuthorizationClient authorizationClient,
     ILogger<SigningService> logger,
     ISigneeContextsManager signeeContextsManager,
+    ISignDocumentManager signDocumentManager,
     Telemetry? telemetry = null
 ) : ISigningService
 {
@@ -44,6 +42,7 @@ internal sealed class SigningService(
     );
     private readonly ILogger<SigningService> _logger = logger;
     private readonly ISigneeContextsManager _signeeContextsManager = signeeContextsManager;
+    private readonly ISignDocumentManager _signDocumentManager = signDocumentManager;
     private readonly IAppMetadata _appMetadata = appMetadata;
     private readonly ISigningCallToActionService _signingCallToActionService = signingCallToActionService;
     private const string ApplicationJsonContentType = "application/json";
@@ -155,9 +154,12 @@ internal sealed class SigningService(
 
         var taskId = instanceDataAccessor.Instance.Process.CurrentTask.ElementId;
 
-        List<SignDocument> signDocuments = await DownloadSignDocuments(instanceDataAccessor, signatureConfiguration);
+        List<SignDocument> signDocuments = await _signDocumentManager.GetSignDocuments(
+            instanceDataAccessor,
+            signatureConfiguration
+        );
 
-        await SynchronizeSigneeContextsWithSignDocuments(taskId, signeeContexts, signDocuments);
+        await _signDocumentManager.SynchronizeSigneeContextsWithSignDocuments(taskId, signeeContexts, signDocuments);
 
         return signeeContexts;
     }
@@ -179,9 +181,10 @@ internal sealed class SigningService(
 
         List<string> keyRoleOrganizations = await authorizationClient.GetKeyRoleOrganisationParties(userId, orgNumbers);
 
-        List<OrganizationSignee> authorizedOrganizations = orgSignees
-            .Where(organizationSignee => keyRoleOrganizations.Contains(organizationSignee.OrgNumber))
-            .ToList();
+        List<OrganizationSignee> authorizedOrganizations =
+        [
+            .. orgSignees.Where(organizationSignee => keyRoleOrganizations.Contains(organizationSignee.OrgNumber)),
+        ];
 
         return authorizedOrganizations;
     }
@@ -194,20 +197,15 @@ internal sealed class SigningService(
         CancellationToken ct
     )
     {
-        if (signatureConfiguration.SigneeStatesDataTypeId is not null)
-        {
-            RemoveSigneeState(instanceDataMutator, signatureConfiguration.SigneeStatesDataTypeId);
-        }
-
-        if (signatureConfiguration.SignatureDataType is not null)
-        {
-            RemoveAllSignatures(instanceDataMutator, signatureConfiguration.SignatureDataType);
-        }
+        // cleanup
+        RemoveSigneeState(instanceDataMutator, signatureConfiguration.SigneeStatesDataTypeId);
+        RemoveAllSignatures(instanceDataMutator, signatureConfiguration.SignatureDataType);
 
         List<SigneeContext> signeeContexts = await GetSigneeContexts(instanceDataMutator, signatureConfiguration);
-        List<SigneeContext> signeeContextsWithDelegation = signeeContexts
-            .Where(x => x.SigneeState.IsAccessDelegated)
-            .ToList();
+        List<SigneeContext> signeeContextsWithDelegation =
+        [
+            .. signeeContexts.Where(x => x.SigneeState.IsAccessDelegated),
+        ];
 
         if (signeeContextsWithDelegation.IsNullOrEmpty())
         {
@@ -234,182 +232,6 @@ internal sealed class SigningService(
             signeeContextsWithDelegation,
             ct
         );
-    }
-
-    private async Task<List<SignDocument>> DownloadSignDocuments(
-        IInstanceDataAccessor instanceDataAccessor,
-        AltinnSignatureConfiguration signatureConfiguration
-    )
-    {
-        string signatureDataTypeId =
-            signatureConfiguration.SignatureDataType
-            ?? throw new ApplicationConfigException("SignatureDataType is not set in the signature configuration.");
-
-        IEnumerable<DataElement> signatureDataElements = instanceDataAccessor.GetDataElementsForType(
-            signatureDataTypeId
-        );
-
-        try
-        {
-            SignDocument[] signDocuments = await Task.WhenAll(
-                signatureDataElements.Select(signatureDataElement =>
-                    DownloadSignDocumentAsync(instanceDataAccessor, signatureDataElement)
-                )
-            );
-
-            return [.. signDocuments];
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to download signature documents.");
-            throw;
-        }
-    }
-
-    private async Task<SignDocument> DownloadSignDocumentAsync(
-        IInstanceDataAccessor instanceDataAccessor,
-        DataElement signatureDataElement
-    )
-    {
-        try
-        {
-            ReadOnlyMemory<byte> data = await instanceDataAccessor.GetBinaryData(signatureDataElement);
-            string signDocumentSerialized = Encoding.UTF8.GetString(data.ToArray());
-
-            return JsonSerializer.Deserialize<SignDocument>(signDocumentSerialized, _jsonSerializerOptions)
-                ?? throw new JsonException("Could not deserialize signature document.");
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(
-                ex,
-                "Failed to download signature document for DataElement with ID {DataElementId}.",
-                signatureDataElement.Id
-            );
-            throw;
-        }
-    }
-
-    /// <summary>
-    /// This method exists to ensure we have a SigneeContext for both signees that have been delegated access to sign and signees that have signed using access granted through the policy.xml file.
-    /// </summary>
-    internal async Task SynchronizeSigneeContextsWithSignDocuments(
-        string taskId,
-        List<SigneeContext> signeeContexts,
-        List<SignDocument> signDocuments
-    )
-    {
-        _logger.LogDebug(
-            "Synchronizing signee contexts {SigneeContexts} with sign documents {SignDocuments} for task {TaskId}.",
-            JsonSerializer.Serialize(signeeContexts, _jsonSerializerOptions),
-            JsonSerializer.Serialize(signDocuments, _jsonSerializerOptions),
-            taskId
-        );
-
-        List<SignDocument> unmatchedSignDocuments = signDocuments;
-
-        // OrganizationSignee is most general, so it should be sorted to the end of the list
-        signeeContexts.Sort(
-            (a, b) =>
-                a.Signee is OrganizationSignee ? 1
-                : b.Signee is OrganizationSignee ? -1
-                : 0
-        );
-
-        foreach (SigneeContext signeeContext in signeeContexts)
-        {
-            SignDocument? matchedSignDocument = signDocuments.FirstOrDefault(signDocument =>
-            {
-                return signeeContext.Signee switch
-                {
-                    PersonSignee personSignee => IsPersonSignDocument(signDocument)
-                        && personSignee.SocialSecurityNumber == signDocument.SigneeInfo.PersonNumber,
-                    PersonOnBehalfOfOrgSignee personOnBehalfOfOrgSignee => IsPersonOnBehalfOfOrgSignDocument(
-                        signDocument
-                    )
-                        && personOnBehalfOfOrgSignee.OnBehalfOfOrg.OrgNumber
-                            == signDocument.SigneeInfo.OrganisationNumber
-                        && personOnBehalfOfOrgSignee.SocialSecurityNumber == signDocument.SigneeInfo.PersonNumber,
-                    SystemSignee systemSignee => IsSystemSignDocument(signDocument)
-                        && systemSignee.OnBehalfOfOrg.OrgNumber == signDocument.SigneeInfo.OrganisationNumber
-                        && systemSignee.SystemId.Equals(signDocument.SigneeInfo.SystemUserId),
-                    OrganizationSignee orgSignee => IsOrgSignDocument(signDocument)
-                        && orgSignee.OrgNumber == signDocument.SigneeInfo.OrganisationNumber,
-
-                    _ => throw new InvalidOperationException("Signee is not of a supported type."),
-                };
-            });
-
-            if (matchedSignDocument is not null)
-            {
-                if (signeeContext.Signee is OrganizationSignee orgSignee)
-                {
-                    await ConvertOrgSignee(matchedSignDocument, signeeContext, orgSignee);
-                }
-
-                signeeContext.SignDocument = matchedSignDocument;
-                unmatchedSignDocuments.Remove(matchedSignDocument);
-            }
-        }
-
-        // Create new contexts for documents that aren't matched with existing signee contexts
-        foreach (SignDocument signDocument in unmatchedSignDocuments)
-        {
-            SigneeContext newSigneeContext = await CreateSigneeContextFromSignDocument(taskId, signDocument);
-            signeeContexts.Add(newSigneeContext);
-        }
-    }
-
-    private async Task ConvertOrgSignee(
-        SignDocument? signDocument,
-        SigneeContext orgSigneeContext,
-        OrganizationSignee orgSignee
-    )
-    {
-        if (signDocument is null)
-        {
-            return;
-        }
-
-        var signeeInfo = signDocument.SigneeInfo;
-
-        if (!string.IsNullOrEmpty(signeeInfo.PersonNumber))
-        {
-            orgSigneeContext.Signee = await orgSignee.ToPersonOnBehalfOfOrgSignee(
-                signeeInfo.PersonNumber,
-                altinnPartyClient.LookupParty
-            );
-        }
-        else if (signeeInfo.SystemUserId.HasValue)
-        {
-            orgSigneeContext.Signee = orgSignee.ToSystemSignee(signeeInfo.SystemUserId.Value);
-        }
-        else
-        {
-            throw new InvalidOperationException("Signee is neither a person nor a system user");
-        }
-    }
-
-    private async Task<SigneeContext> CreateSigneeContextFromSignDocument(string taskId, SignDocument signDocument)
-    {
-        _logger.LogDebug(
-            "Creating signee context for sign document {SignDocument} for task {TaskId}.",
-            JsonSerializer.Serialize(signDocument, _jsonSerializerOptions),
-            taskId
-        );
-
-        return new SigneeContext
-        {
-            TaskId = taskId,
-            Signee = await From(
-                signDocument.SigneeInfo.PersonNumber,
-                signDocument.SigneeInfo.OrganisationNumber,
-                signDocument.SigneeInfo.SystemUserId,
-                LookupParty
-            ),
-            SigneeState = new SigneeState() { IsAccessDelegated = true, HasBeenMessagedForCallToSign = true },
-            SignDocument = signDocument,
-        };
     }
 
     private async Task<Party> LookupParty(PartyLookup partyLookup)
@@ -459,9 +281,7 @@ internal sealed class SigningService(
                 serviceOwnerDetails.Orgnr = "991825827";
             }
 
-            serviceOwnerParty = await altinnPartyClient.LookupParty(
-                new PartyLookup { OrgNo = serviceOwnerDetails?.Orgnr }
-            );
+            serviceOwnerParty = await LookupParty(new PartyLookup { OrgNo = serviceOwnerDetails?.Orgnr });
         }
         catch (Exception e)
         {
@@ -472,8 +292,13 @@ internal sealed class SigningService(
         return (serviceOwnerParty, true);
     }
 
-    private void RemoveSigneeState(IInstanceDataMutator instanceDataMutator, string signeeStatesDataTypeId)
+    private void RemoveSigneeState(IInstanceDataMutator instanceDataMutator, string? signeeStatesDataTypeId)
     {
+        if (string.IsNullOrEmpty(signeeStatesDataTypeId))
+        {
+            return;
+        }
+
         using Activity? activity = telemetry?.StartRemoveSigneeStateActivity();
 
         IEnumerable<DataElement> signeeStateDataElements = instanceDataMutator.GetDataElementsForType(
@@ -489,6 +314,11 @@ internal sealed class SigningService(
 
     private void RemoveAllSignatures(IInstanceDataMutator instanceDataMutator, string signatureDataType)
     {
+        if (string.IsNullOrEmpty(signatureDataType))
+        {
+            return;
+        }
+
         using Activity? activity = telemetry?.StartRemoveAllSignaturesActivity();
 
         IEnumerable<DataElement> signatures = instanceDataMutator.GetDataElementsForType(signatureDataType);
@@ -497,28 +327,5 @@ internal sealed class SigningService(
         {
             instanceDataMutator.RemoveDataElement(signature);
         }
-    }
-
-    private static bool IsPersonOnBehalfOfOrgSignDocument(SignDocument signDocument)
-    {
-        return !string.IsNullOrEmpty(signDocument.SigneeInfo.PersonNumber)
-            && !string.IsNullOrEmpty(signDocument.SigneeInfo.OrganisationNumber);
-    }
-
-    private static bool IsPersonSignDocument(SignDocument signDocument)
-    {
-        return !string.IsNullOrEmpty(signDocument.SigneeInfo.PersonNumber)
-            && string.IsNullOrEmpty(signDocument.SigneeInfo.OrganisationNumber);
-    }
-
-    private static bool IsOrgSignDocument(SignDocument signDocument)
-    {
-        return !string.IsNullOrEmpty(signDocument.SigneeInfo.OrganisationNumber);
-    }
-
-    private static bool IsSystemSignDocument(SignDocument signDocument)
-    {
-        return !string.IsNullOrEmpty(signDocument.SigneeInfo.OrganisationNumber)
-            && signDocument.SigneeInfo.SystemUserId.HasValue;
     }
 }
