+++ src/Altinn.App.Core/Features/Telemetry/Telemetry.SigningService.cs
@@ -76,6 +76,9 @@ partial class Telemetry
     internal Activity? StartRemoveAllSignaturesActivity() =>
         ActivitySource.StartActivity("SigningService.RemoveAllSignatures");
 
+    internal Activity? StartGetSignDocumentsActivity() =>
+        ActivitySource.StartActivity("SigningService.GetSignDocuments");
+
     internal Activity? StartSignActivity() => ActivitySource.StartActivity("SigningService.Sign"); // TODO: expand to include signee id
 
     internal static class DelegationConst
